requirements.txt


# Django
Django==1.11
djangorestframework==3.6.3
django-cors-headers
mysqlclient==1.3.10
Pillow==4.2.1
coreapi-cli
Markdown
Pygments
pytz


# Authentication
djangorestframework-jwt==1.11.0


# Payment
zeep
pytz
oauth2client
pycrypto == 2.6.1


# Third party
requests==2.18.3
ipython

